Renamed plugins to features
