+++ src/Http/MockResponse.php
@@ -6,7 +6,7 @@ use GuzzleHttp\Psr7\Response;
 use Sammyjo20\Saloon\Traits\CollectsData;
 use Sammyjo20\Saloon\Traits\CollectsConfig;
 use Sammyjo20\Saloon\Traits\CollectsHeaders;
-use Sammyjo20\Saloon\Traits\Features\HasBody;
+use Sammyjo20\Saloon\Traits\Plugins\HasBody;
 
 class MockResponse
 {
