--- src/Traits/Features/AcceptsJson.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Sammyjo20\Saloon\Traits\Features;
-
-trait AcceptsJson
-{
-    public function bootAcceptsJson()
-    {
-        $this->addHeader('Accept', 'application/json');
-    }
-}
