+++ tests/Resources/Connectors/InterceptedConnector.php
@@ -3,7 +3,7 @@
 namespace Sammyjo20\Saloon\Tests\Resources\Connectors;
 
 use Sammyjo20\Saloon\Http\SaloonConnector;
-use Sammyjo20\Saloon\Traits\Features\AcceptsJson;
+use Sammyjo20\Saloon\Traits\Plugins\AcceptsJson;
 
 class InterceptedConnector extends SaloonConnector
 {
