+++ tests/Resources/Connectors/InvalidDefinedRequestSelectionConnector.php
@@ -4,7 +4,7 @@ namespace Sammyjo20\Saloon\Tests\Resources\Connectors;
 
 use Sammyjo20\MissingClass;
 use Sammyjo20\Saloon\Http\SaloonConnector;
-use Sammyjo20\Saloon\Traits\Features\AcceptsJson;
+use Sammyjo20\Saloon\Traits\Plugins\AcceptsJson;
 use Sammyjo20\Saloon\Tests\Resources\Requests\UserRequest;
 use Sammyjo20\Saloon\Tests\Resources\Requests\ErrorRequest;
