+++ tests/Resources/Connectors/WithBootConnector.php
@@ -4,7 +4,7 @@ namespace Sammyjo20\Saloon\Tests\Resources\Connectors;
 
 use Sammyjo20\Saloon\Http\SaloonRequest;
 use Sammyjo20\Saloon\Http\SaloonConnector;
-use Sammyjo20\Saloon\Traits\Features\AcceptsJson;
+use Sammyjo20\Saloon\Traits\Plugins\AcceptsJson;
 
 class WithBootConnector extends SaloonConnector
 {
