+++ tests/Resources/Requests/AlwaysThrowRequest.php
@@ -4,7 +4,7 @@ namespace Sammyjo20\Saloon\Tests\Resources\Requests;
 
 use Sammyjo20\Saloon\Constants\Saloon;
 use Sammyjo20\Saloon\Http\SaloonRequest;
-use Sammyjo20\Saloon\Traits\Features\AlwaysThrowsOnErrors;
+use Sammyjo20\Saloon\Traits\Plugins\AlwaysThrowsOnErrors;
 use Sammyjo20\Saloon\Tests\Resources\Connectors\TestConnector;
 
 class AlwaysThrowRequest extends SaloonRequest
