+++ tests/Resources/Requests/CachedUserRequest.php
@@ -4,7 +4,7 @@ namespace Sammyjo20\Saloon\Tests\Resources\Requests;
 
 use Sammyjo20\Saloon\Constants\Saloon;
 use Sammyjo20\Saloon\Http\SaloonRequest;
-use Sammyjo20\Saloon\Traits\Features\CachesRequests;
+use Sammyjo20\Saloon\Traits\Plugins\CachesRequests;
 use Sammyjo20\Saloon\Tests\Resources\Connectors\TestConnector;
 
 class CachedUserRequest extends SaloonRequest
