+++ tests/Unit/Plugins/AlwaysThrowOnErrorsFeatureTest.php
