+++ tests/Unit/Plugins/HasBodyFeatureTest.php
