+++ tests/Unit/Plugins/HasXMLBodyFeatureTest.php
